+++ packages/glob/RELEASES.md
@@ -1,9 +1,13 @@
 # @actions/glob Releases
 
-### 0.1.0
+### 0.1.2
 
-- Initial release
+- [Fix bug where files were matched incorrectly](https://github.com/actions/toolkit/pull/805)
 
 ### 0.1.1
 
 - Update @actions/core version
+### 0.1.0
+
+- Initial release
+